directory renaming
the now common_methods-directory had not matched the new naming-style
and addionally the data_structure was renamed to common_items as preparation
for the next add new table-items.
