+++ include/libKitsuneCommon/common_methods/string_methods.h
