+++ include/libKitsuneCommon/common_methods/vector_methods.h
