+++ src/src.pro
@@ -13,14 +13,14 @@ SOURCES += \
     threading/thread.cpp \
     statemachine/statemachine.cpp \
     buffering/data_buffer_methods.cpp \
-    data_structure/data_items.cpp
+    common_items/data_items.cpp
 
 HEADERS += \
     ../include/libKitsuneCommon/buffering/data_buffer.h \
     ../include/libKitsuneCommon/buffering/data_buffer_methods.h \
-    ../include/libKitsuneCommon/commonMethods/string_methods.h \
-    ../include/libKitsuneCommon/commonMethods/vector_methods.h \
-    ../include/libKitsuneCommon/data_structure/data_items.h \
+    ../include/libKitsuneCommon/common_methods/string_methods.h \
+    ../include/libKitsuneCommon/common_methods/vector_methods.h \
+    ../include/libKitsuneCommon/common_items/data_items.h \
     ../include/libKitsuneCommon/statemachine/statemachine.h \
     ../include/libKitsuneCommon/testing/test.h \
     ../include/libKitsuneCommon/threading/thread.h \
