+++ tests/libKitsuneCommon/common_items/data_items_DataObject_test.cpp
@@ -8,7 +8,7 @@
  */
 
 #include "data_items_DataObject_test.h"
-#include <data_structure/data_items.h>
+#include <common_items/data_items.h>
 
 namespace Kitsune
 {
